Create businessLayer.py
This is a complete haul from the procedural version of the filtering image program to OO.  Currently working on adding the data level where data entered can be stored. Also working on making the gui that appears more appealing but this is a working OO version of the filtering image program with GUI as the way to enter, resize, filter, view, save images.

There is improper naming conventions which will be fixed prior to submission. 

The commented out sections of code are to be kept as they may be leveraged later on. As it stands I do not know how.