History feature & Error alerting
The goal of this iteration for me was to add the ability to view the history of the program runs. This was done using ttk trees to mimic a table. I also enforced a header to be entered into the beginning of the csv at the creation of the csv. 

I added additional error handling on possible errors that could be thrown. I then ensured proper valid data would be entered. 

I then made it so windows will be removed from view when a file is loaded into memory.and when a file is saved - but only when successful. It will also disappear when history is selected. 

Cheers
SP