Logic Bug Fix
Do to what I believe was a logic error when the program would run it would extend the list on top of the first entered image path. To ensure proper records are made of each entry the clearing of the Global List that acts as the intermediary is cleared at the end of each writing of the csv.  

@line 21

-SP